Invoice saved!
It saves the invoice: some general data extracted from the signed and
verified Facturae file coming from the app mobile, the IV and simkey
used to encrypt part of the data (seller taxInformationNumber and
invoiceNumber), and the Facturae file encrypted.

The IV and simKey are encrypted with PKI.

Next step will be download an indivudual invoice.